address highlighting for events with background colors
Events with background colors had their highlighting behavior disabled, because since an event has to be selected to add a color, users thought the color adding didn't work.  However it was pointed out in #4129 that users still would like to know which event is selected.  So this PR attempts to solve the problem by drawing a border over a selected event if that event has a background color.