Fix Texture Replace and Team Colors with Ship Changes (#4207)
* Fix Texture Replace with Ship Changes

Tabled `$Texture Replace` did not work on ships where the class is changed in the loadout or with the `change-ship-class` sexp. This was because the replacement textures only were loaded during mission parse, and the `change_ship_type` did not account for loading new textures from ship classes that were not in the initial mission parse. Part of the bug was also that the code that updates texture replacements was after the texture page-in section of `change_ship_type`.

This PR fixes that bug by properly loading the textures in when the ship class changes. It also moves the code to within the `ship_model_change` function (which is only used in the `change_ship_type` function) so that the texture loading happens correctly before texture page in.

* fix accidental whitespace changes

* fix comment

* also fix team colors

* comment spelling fix

* appease clang

* properly clear replace textures between ship changes