convert strncpy to strncpy_s
I have gone through the codebase and replaced most instances of `strncpy` with `strncpy_s` which always adds a null-terminator and includes some extra error checking.  Quite a few of these places did not use the length properly and have been fixed.

This is based on #4321 but without the safe_strings removal.