Allows default shadow quality to be per mod
This PR allows mods to set a default `Shadow_detail` level in `game_settings.tbl`. ATM, if the `enable_shadows` flag is set then the game will use the `Medium` level.

It is very useful for mods to be able to specify a default level. For example, with FotG we have noticed that if the `High` shadows level looks far, far better then the default of `Medium`, with getting no noticable FSP difference on our testers range of dedicated graphics cards. Yes, using `shadow_quality` flag will work, but that's a tedious extra step, especially for our testers who are new to FSO. Plus, using a `shadow_qualilty` that overrides if `enable_shadows` is set or not is rather unintuitive for our players.

A boolean check is used because it's far easier to set and check in the mod table code, especially for default values of `false`.