Allow tech-room to auto switch to singleplayer
Many buttons in the main menu auto switch to either multiplayer mode or singleplayer mode. For example, the campaign door always switches to singleplayer, while the multiplayer door auto switches to multiplayer mode.

This is very useful for mods, especially with new players who are not sure where and how to switch between multi and single player pilots.

This PR adds an option to allow mods to auto switch to singleplayer for the techroom, which is especially useful for playing simulator missions. The method used here is the same method that the campaign door uses to auto switch to singleplayer.

Tested and works as expected.