remove the special_ship field
This field existed in retail, but was not used for its intended purpose.  Subsequently, I modified the functionality so that it worked as advertised, but this has side-effects for retail missions which specify a special ship other than the wing leader.  I'm not sure to what extent this field actually enhances a mission, so I propose removing it.