add some XSTR entries
Add some entries for "or" and "None", mostly in the controls code.  Note that there is already an XSTR entry for "none", but it does not have the right capitalization.  Also note that "None" is localized in `textify_scancode` but not `textify_scancode_universal`.

This also makes "Nothing" cargo use appropriate localization.

Fixes #5227.